+++ views/default/forms/messages/send.php
@@ -7,32 +7,23 @@
  */
 
 blockuser_access_messages('/blockuser', get_input('send_to'));
-$recipient_guid = elgg_extract('recipient_guid', $vars, 0);
+
+$recipient_username = elgg_extract('recipient_username', $vars, '');
 $subject = elgg_extract('subject', $vars, '');
 $body = elgg_extract('body', $vars, '');
 
-$recipients_options = array();
-foreach ($vars['friends'] as $friend) {
-	$recipients_options[$friend->guid] = $friend->name;
-}
-
-if (!array_key_exists($recipient_guid, $recipients_options)) {
-	$recipient = get_entity($recipient_guid);
-	if (elgg_instanceof($recipient, 'user')) {
-		$recipients_options[$recipient_guid] = $recipient->name;
-	}
-}
-
-$recipient_drop_down = elgg_view('input/dropdown', array(
-	'name' => 'recipient_guid',
-	'value' => $recipient_guid,
-	'options_values' => $recipients_options,
+$recipient_autocomplete = elgg_view('input/autocomplete', array(
+	'name' => 'recipient_username',
+	'value' => $recipient_username,
+	'match_on' => array('friends'),
 ));
 
 ?>
 <div>
-	<label><?php echo elgg_echo("messages:to"); ?>: </label>
-	<?php echo $recipient_drop_down; ?>
+	<label><?php echo elgg_echo("email:to"); ?>: </label>
+	<?php echo $recipient_autocomplete; ?>
+	<span class="elgg-text-help"><?php echo elgg_echo("messages:to:help"); ?></span>
+	
 </div>
 <div>
 	<label><?php echo elgg_echo("messages:title"); ?>: <br /></label>
@@ -51,5 +42,5 @@ $recipient_drop_down = elgg_view('input/dropdown', array(
 	?>
 </div>
 <div class="elgg-foot">
-	<?php echo elgg_view('input/submit', array('value' => elgg_echo('messages:send'))); ?>
+	<?php echo elgg_view('input/submit', array('value' => elgg_echo('send'))); ?>
 </div>